+++ src/backend/langboard/core/bootstrap/uvicorn/supervisors/multiprocess.py
@@ -159,18 +159,23 @@ class Multiprocess:
             process = Process(self.config, self.target, self.sockets)
             process.start()
             self.processes.append(process)
-        self.broker_process = run_broker(is_restarting=False)
+        if getattr(self.config, "enable_broker", True):
+            self.broker_process = run_broker(is_restarting=False)
+        else:
+            self.broker_process = None
 
     def terminate_all(self) -> None:
         for process in self.processes:
             process.terminate()
-        self.broker_process.terminate()
-        self.broker_process.kill()
+        if self.broker_process:
+            self.broker_process.terminate()
+            self.broker_process.kill()
 
     def join_all(self) -> None:
         for process in self.processes:
             process.join()
-        self.broker_process.join()
+        if self.broker_process:
+            self.broker_process.join()
 
     def restart_all(self) -> None:
         for idx, process in enumerate(self.processes):
@@ -179,10 +184,14 @@ class Multiprocess:
             new_process = Process(self.config, self.target, self.sockets)
             new_process.start()
             self.processes[idx] = new_process
-        self.broker_process.terminate()
-        self.broker_process.kill()
-        self.broker_process.join()
-        self.broker_process = run_broker(is_restarting=True)
+        if self.broker_process:
+            self.broker_process.terminate()
+            self.broker_process.kill()
+            self.broker_process.join()
+            if getattr(self.config, "enable_broker", True):
+                self.broker_process = run_broker(is_restarting=True)
+            else:
+                self.broker_process = None
 
     def run(self) -> None:
         message = f"Started parent process [{os.getpid()}]"
@@ -197,6 +206,7 @@ class Multiprocess:
 
         self.terminate_all()
         self.join_all()
+        DispatcherQueue.close()
 
         message = f"Stopping parent process [{os.getpid()}]"
         color_message = "Stopping parent process [{}]".format(click.style(str(os.getpid()), fg="cyan", bold=True))
@@ -221,11 +231,14 @@ class Multiprocess:
             process.start()
             self.processes[idx] = process
 
-        if not self.broker_process.is_alive():
+        if self.broker_process and not self.broker_process.is_alive():
             self.broker_process.terminate()
             self.broker_process.kill()
             self.broker_process.join()
-            self.broker_process = run_broker(is_restarting=True)
+            if getattr(self.config, "enable_broker", True):
+                self.broker_process = run_broker(is_restarting=True)
+            else:
+                self.broker_process = None
 
     def handle_signals(self) -> None:
         for sig in tuple(self.signal_queue):
